requirements.txt
traitlets==4.3.3
matplotlib==3.2.0
torch==1.4.0
pandas==1.0.1
numpy==1.18.1
torchvision==0.5.0
scikit_learn==0.22.2.post1
sty
munch
wandb
shortuuid
seaborn
plotly
pillow
